% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcomment.R
\name{bcomment}
\alias{bcomment}
\title{Wrapper function around \code{colfmt()} to make the font blue.}
\usage{
bcomment(x)
}
\arguments{
\item{x}{character string with comment}
}
\description{
In addition to adding color, underscores are prepended and
append for HTML and \code{\\emph{}} is used for LaTeX
}
