% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_code_lines.R
\name{classify_code_lines}
\alias{classify_code_lines}
\title{Creates a list object containing number of lines, blank lines,
bare comments, roxygen2 documentation, and code.}
\usage{
classify_code_lines(files = ".", path = NULL)
}
\arguments{
\item{files}{character vector of file names. The file names may contain the
path. Without a path in the file name and without a defined \code{path}
paramenter, the current directory is used.}

\item{path}{optional character vector of length one or of the same length as
\code{files}. If of length one, it is used for all files. If it is the
same length as files, each value is paired with its corresponding file to
make the fully qualified name.}
}
\description{
Creates a list object containing number of lines, blank lines,
bare comments, roxygen2 documentation, and code.
}
