% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearpage.R
\name{clearpage}
\alias{clearpage}
\title{clearpage that works for PDF (LaTeX) and HTML output from \code{knitr}}
\usage{
clearpage()
}
\value{
Returns either a LaTeX command, \code{\\clearpage}, if 
\code{is_latex_output()} is \code{TRUE}, an HTML command,
\code{<div style="page-break-before: always;" />} if 
\code{is_html_output()} is \code{TRUE}, and \code{NULL} if 
neither is \code{TRUE}.
}
\description{
Causes the print buffer to be flushed.
}
\examples{
clearbreak()
}
