% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colfmt.R
\name{colfmt}
\alias{colfmt}
\title{Returns a string formated as a comment for HTML or LaTeX depending on
output format.}
\usage{
colfmt(x, color = "red")
}
\arguments{
\item{x}{character string with comment.}

\item{color}{color that font is to be}
}
\description{
Returns a string formated as a comment for HTML or LaTeX depending on
output format.
}
