% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_vectors.R
\name{combine_vectors}
\alias{combine_vectors}
\title{Returns a character vector made up of two vectors, where if both are is.na()
NA is returned, if one is.na() the other is returned and if neither are
is.na() then the first is returned.}
\usage{
combine_vectors(vec1, vec2)
}
\arguments{
\item{vec1}{first character vector}

\item{vec2}{second character vector}
}
\description{
Returns a character vector made up of two vectors, where if both are is.na()
NA is returned, if one is.na() the other is returned and if neither are
is.na() then the first is returned.
}
