% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_selected_comments_from_lines.R
\name{delete_selected_comments_from_lines}
\alias{delete_selected_comments_from_lines}
\title{Delete selected comments from lines.}
\usage{
delete_selected_comments_from_lines(lines, label = "", isolated = TRUE)
}
\arguments{
\item{lines}{Character vector of lines to examine.}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding each element of the character vector in turn 
immediately after "<!--" in the regex expression. The resulting logical
vectors are OR'd together to combine their results.}

\item{isolated}{Logical value indicating that the comment is the first text
to appear on the line.}
}
\description{
Delete selected comments from lines.
}
