% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_selected_code_comments.R
\name{display_selected_code_comments}
\alias{display_selected_code_comments}
\title{Displays selected comments}
\usage{
display_selected_code_comments(path = ".", pattern = NULL, label = "")
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to 
the working directory, getwd(). Tilde expansion (see path.expand) is 
performed. Missing values will be ignored. 
Elements with a marked encoding will be converted to the native encoding 
(and if that fails, considered non-existent). Defaults to ".".}

\item{pattern}{an optional regular expression. Only file names which match the 
regular expression will be returned.}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding each element of the character vector in turn 
immediately after "<!--" in the regex expression. The resulting logical
vectors are OR'd together to combine their results.}
}
\value{
Dataframe of selected comments with the base file name, the 
comment label, the comment start line, and the comment text.

Internally uses the \code{list.files} function with the \code{path} and 
\code{pattern} arguments as defined in the call. Other arguments to
\code{list.files} are forced as follows:
\describe{
\item{all.files}{TRUE}
\item{full.names}{TRUE}
\item{recursive}{TRUE}
\item{ignore.case}{TRUE}
\item{include.dirs}{FALSE}
\item{no..}{FALSE}
}
The user is free to create the list of files anyway desired and provide them
to the \code{path} argument.
}
\description{
Displays selected comments
}
