% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups_to_empty_str.R
\name{dups_to_empty_str}
\alias{dups_to_empty_str}
\title{Replaces duplicate strings with empty strings.}
\usage{
dups_to_empty_str(char_vec)
}
\arguments{
\item{char_vec}{character vector}
}
\description{
Replaces duplicate strings with empty strings
in a character vector.
}
\examples{
library(stringi)
dups_to_empty_str(c("A", "duplicated", "duplicated", "word"))
}
