% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expunge_names.R
\name{expunge_names}
\alias{expunge_names}
\title{Returns a character vector with the string pattern
within the expunge_str removed from names_str if they existed.}
\usage{
expunge_names(names_str, expunge_str = c(), lower = TRUE)
}
\arguments{
\item{names_str}{character vector with names to be filtered}

\item{expunge_str}{character vector with names to be expunged}

\item{lower}{logical TRUE if names_str is to be turned to lower case}
}
\description{
Returns a character vector with the string pattern
within the expunge_str removed from names_str if they existed.
}
