% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_html_comments.R
\name{find_html_comments}
\alias{find_html_comments}
\title{Find HTML comment}
\usage{
find_html_comments(lines, label = "")
}
\arguments{
\item{lines}{Character vector of lines to examine.}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding this character vector of length one immediately 
after "<!--".}
}
\description{
Find HTML comment
}
