% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_age}
\alias{get_age}
\title{Returns the fractional age in years, months or days between the start_date
and the end_date.}
\usage{
get_age(start_date, end_date, unit)
}
\arguments{
\item{start_date}{a POSIXct or POSIXt object with the beginning date (such
as a birth date).}

\item{end_date}{a POSIXct or POSIXt object with the ending date (such
as a death date).}

\item{unit}{one of c("year", "month", "week", "day") that become the
unit used as the devisor for the time interval.}
}
\description{
Returns the fractional age in years, months or days between the start_date
and the end_date.
}
