% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_col_width_fmt}
\alias{get_col_width_fmt}
\title{Returns a format statement used by xtable based on centimeter column widths}
\usage{
get_col_width_fmt(col_widths_txt)
}
\arguments{
\item{col_widths_txt}{character vector with the representations of the
column widths to be used in the format statement.}
}
\description{
Returns a format statement used by xtable based on centimeter column widths
}
