% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dated_excel_name.R
\name{get_dated_excel_name}
\alias{get_dated_excel_name}
\title{Returns a character vector with an Excel complient name having the date prepended.}
\usage{
get_dated_excel_name(data_set_name)
}
\arguments{
\item{data_set_name}{character vector with name to use in file name}
}
\description{
Returns a character vector with an Excel complient name having the date prepended.
}
