% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_dated_filename}
\alias{get_dated_filename}
\title{Returns a character vector with an file name having the date prepended.}
\usage{
get_dated_filename(filename)
}
\arguments{
\item{filename}{character vector with name to use in file name}
}
\description{
Returns a character vector with an file name having the date prepended.
}
