% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_lines.R
\name{get_example_lines}
\alias{get_example_lines}
\title{Get the examples lines from a source code file.}
\usage{
get_example_lines(file = stdin())
}
\arguments{
\item{file}{character vector of length 1 with the file to extract example
code from.}
}
\value{
Character vector with one lines from the 
\code{examples} section of the code.
}
\description{
User provides the source code file path. This a a simplistic
function that simply returns lines between \code{@examples} and the next
line that starts with \code{#' @}. This is not robust and is completely 
dependent on my coding style.
}
