% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_and_function_names.R
\name{get_files_and_function_names}
\alias{get_files_and_function_names}
\title{Get the file names and functions defined in those files}
\usage{
get_files_and_function_names(path = ".")
}
\arguments{
\item{path}{character vector of length one having the directory path to
the files containing function definitions.}
}
\value{
data.frame with a column of file names and the corresponding
`function\ *(` and must be the first non-blank token on a line.
}
\description{
Get the file names and functions defined in those files
}
