% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_json_message.R
\name{get_first_json_message}
\alias{get_first_json_message}
\title{Returns a list containing two objects in the text of a character vector
of length one: (1) object = the first json object found and (2) remainder =
the remaining text.}
\usage{
get_first_json_message(json_txt)
}
\arguments{
\item{json_txt}{character vector of length one having one or more JSON
objects in character form.}
}
\description{
Properly formed messages are assumed. Error checking is non-existent.
}
