% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_names.R
\name{get_function_names}
\alias{get_function_names}
\title{Get functions defined in a vector of lines}
\usage{
get_function_names(lines)
}
\arguments{
\item{lines}{character vector with text from file}
}
\value{
A character vector of function names or an empty character vector
}
\description{
Ignores all comment lines and defines a function name to be
the set of contiguous (non-blank characters) immediately prior to
`function\ *(` and must be the first non-blank token on a line.
}
