% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_html_comment_text_lines_and_labels_from_files.R
\name{get_html_comment_text_lines_and_labels_from_files}
\alias{get_html_comment_text_lines_and_labels_from_files}
\title{Read in file and find, classify, and count comments by label type.}
\usage{
get_html_comment_text_lines_and_labels_from_files(files, label = "")
}
\arguments{
\item{files}{Names of files}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding each element of the character vector in turn 
immediately after "<!--" in the regex expression. The resulting logical
vectors are OR'd together to combine their results.}
}
\description{
Read in file and find, classify, and count comments by label type.
}
