% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_json_list}
\alias{get_json_list}
\title{Returns list of lists made by call to fromJSON()}
\usage{
get_json_list(json_txt)
}
\arguments{
\item{json_txt}{character vector of length 1 having one or more
JSON objects in text form.}
}
\description{
Returns list of lists made by call to fromJSON()
}
