% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_functions_in_examples.R
\name{get_pkg_functions_in_examples}
\alias{get_pkg_functions_in_examples}
\title{Get a data.frame with a list of package functions used within the 
\code{examples} sections of the source code.}
\usage{
get_pkg_functions_in_examples(path = ".")
}
\arguments{
\item{path}{character vector of length one having the full path name.}
}
\value{
a data.frame containing three columns \code{path} with the full path
name of the source file, \code{file} with the basename of the code file, and
\code{function_names} with the name of the function used in the 
\code{examples} section of the documentation and are also functions 
defined within the package.
}
\description{
Get a data.frame with a list of package functions used within the 
\code{examples} sections of the source code.
}
