% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_list.R
\name{get_pkg_list}
\alias{get_pkg_list}
\title{Provides a sorted CSV list of packages loaded in an R session}
\usage{
get_pkg_list(session_info = sessionInfo(), base = FALSE)
}
\arguments{
\item{session_info}{object returned by \code{sessionInfo}. Defaults to
sessionInfo().}

\item{base}{logical value if TRUE base packages are included}
}
\value{
character string with package names separated by \",\\n\".
}
\description{
Provides a sorted CSV list of packages loaded in an R session
}
\examples{
write.csv(data.frame(packages = get_pkg_list(base = TRUE), stringsAsFactors = FALSE), 
          file = "package_list.csv", row.names = FALSE, quote = FALSE)

}
