% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relative_path_file_names.R
\name{get_relative_path_file_names}
\alias{get_relative_path_file_names}
\title{Creates file names using a common relative path structure}
\usage{
get_relative_path_file_names(files, new_path = ".")
}
\arguments{
\item{files}{character vector of full path file names}

\item{new_path}{character vector of length one having the new file path
common to be prepended to the relative paths created. Default is ".".}
}
\value{
A dataframe with columns `original_file` and `new_file`
}
\description{
Creates file names using a common relative path structure
}
