% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_split_format}
\alias{get_split_format}
\title{Returns a character string of format statements split up from a format
statement based on the number of centimeters indicated by the parameter cm}
\usage{
get_split_format(fmt, cm = 15)
}
\arguments{
\item{fmt}{source format statement}

\item{cm}{number of centimeters format statement are to fit within}
}
\description{
Returns a character string of format statements split up from a format
statement based on the number of centimeters indicated by the parameter cm
}
