% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_table.R
\name{get_split_table}
\alias{get_split_table}
\title{Returns a list of data frames split up for printing so that they do not
exceed the width (cm) specified.}
\usage{
get_split_table(df, fmt, cm = 15)
}
\arguments{
\item{df}{source dataframe}

\item{fmt}{source format statement}

\item{cm}{number of centimeters format statement are to fit within}
}
\description{
Returns a list of data frames split up for printing so that they do not
exceed the width (cm) specified.
}
