% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_files.R
\name{get_target_files}
\alias{get_target_files}
\title{Get a list of files based on presence of a text pattern}
\usage{
get_target_files(path, pattern)
}
\arguments{
\item{path}{A character vector of length one having the path of the directory
to be searched.}

\item{pattern}{A character vector of length one having the pattern to be 
used for matching. Currently only fixed pattern matching is supported.}
}
\value{
A character string containing file paths to targeted files.
}
\description{
Get a list of files based on presence of a text pattern
}
