% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_from_source.R
\name{install_from_source}
\alias{install_from_source}
\title{Install R package from package source}
\usage{
install_from_source(source_names, source_path, install_path)
}
\arguments{
\item{source_names}{character vector having one source package name per
cell.}

\item{source_path}{character vector of length one having the directory
path of where the package sources (*.tar.gz) reside.}

\item{install_path}{character vector of length one having the directory
path of where packages are to be installed.}
}
\description{
Takes a list of packages (\code{source_names}) and the path
(\code{source_path}) to their common location and
installs them into the \code{install_path} directory.
}
