% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{is_even}
\alias{is_even}
\title{Returns even numbers from a set of numbers}
\usage{
is_even(x)
}
\arguments{
\item{x}{vector of numbers (only integers should be used)}
}
\description{
Returns even numbers from a set of numbers
}
