% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{is_file}
\alias{is_file}
\title{Are these files?}
\usage{
is_file(files, no_files_stop = FALSE)
}
\arguments{
\item{files}{character vector of file names or path.}

\item{no_files_stop}{logical vector of length one when set to
\code{TRUE} the routine calls \code{stop} if no files are found.}
}
\description{
If at least one of the character strings is a file it returns
\code{exist} as a list element set to \code{TRUE} and all of the
character strings that are file names are retained and the
character strings that are not file names are removed from the
\code{files} character vector in the returned list.
If none of the  character strings is a file it returns
\code{exist} as a list element set to \code{FALSE}
and all of the character strings are retained in the
\code{files} character vector in the returned list.
}
