% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_braces.R
\name{is_odd}
\alias{is_odd}
\title{Returns odd numbers from a set of numbers}
\usage{
is_odd(x)
}
\arguments{
\item{x}{vector of numbers (only integers should be used)}
}
\description{
Returns odd numbers from a set of numbers
}
