% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_fivenum.R
\name{latex_fivenum}
\alias{latex_fivenum}
\title{Returns character vector of one containing the results of the Tukey
five number summary calculations made by \code{fivenum()}.}
\usage{
latex_fivenum(dist, digits = 3)
}
\arguments{
\item{dist}{vector of numbers that are the argument to \code{fivenum()}}

\item{digits}{integer values with the number of digits to be displayed.}
}
\description{
Returns character vector of one containing the results of the Tukey
five number summary calculations made by \code{fivenum()}.
}
