% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flag.R
\name{make_flag}
\alias{make_flag}
\title{Make a list of functions used flags}
\usage{
make_flag(set_value, flags = c(TRUE, FALSE))
}
\arguments{
\item{set_value}{logical value used to set what is returned}

\item{flags}{vector of possible values for a flag}
}
\value{
A list of two functions: `set` and `value`.
        The function `set` takes a logical variable of length 1 that is used
        to set the logical value returned by `value`.
        The function `value` does not take  an argument and returns either
        the default value provided when the function was created or the 
        last value provided to the `set` function.
}
\description{
Make a list of functions used flags
}
\examples{
# Create function list with `value` defaulted to TRUE
tf_flag <- make_true_false_flag(TRUE)
tf_flag$value()
# reset `value` 
tf_flag$set(FALSE)
tf_flag$value()
# Create function list with `value` defaulted to FALSE
tf_flag <- make_true_false_flag(FALSE)
tf_flag$value()
# reset `value` 
tf_flag$set(TRUE)
tf_flag$value()

}
