% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_license_df.R
\name{make_license_df}
\alias{make_license_df}
\title{Make dataframe with mapping of code license descriptions to more general license descriptions.}
\usage{
make_license_df()
}
\description{
Needs to be manually updated base on descriptions observed.
}
