% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mult_col.R
\name{make_mult_col}
\alias{make_mult_col}
\title{Returns dataframe with multiple sets of the original columns (cols) made
from the original dataframe columns.}
\usage{
make_mult_col(.df, cols = 1)
}
\arguments{
\item{.df}{dataframe}

\item{cols}{number of set of columns to be made}
}
\description{
Thus, a 10 row 1 column dataframe becomes a 4 row 3 column dataframe with
2 NA values added to fill out the end of the new dataframe.
This is used to reduce the length of very narrow dataframes.
}
\examples{
df <- data.frame(a = 1:5, b = letters[1:5], c = LETTERS[1:5])
make_mult_col(df, 2)
}
