% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files_after_deleting_selected_comments.R
\name{make_new_path}
\alias{make_new_path}
\title{Makes the new path if needed}
\usage{
make_new_path(new_path)
}
\arguments{
\item{new_path}{Character vector of length one indicating the directory
newly created files are to be placed.}
}
\value{
NULL

Has the side effect of creating file directories if they do not exists.
It stops the script if it fails.
}
\description{
Makes the new path if needed
}
