% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package_dependency_list.R
\name{make_package_dependency_list}
\alias{make_package_dependency_list}
\title{Make package dependency list}
\usage{
make_package_dependency_list(source_names, path)
}
\arguments{
\item{source_names}{character vector containing the names of the custom
packages to be installed from source.}

\item{path}{character vector of length one having the path to R's
\code{library} directory, which contains the packages being updated from
source. This may or may or may not be the same as the system wide
\code{library} directory. It could be a user directory.}
}
\description{
Gathers the dependencies from each source package and combines the lists
without duplication.
}
