% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pkg_dep_df.R
\name{make_pkg_dep_df}
\alias{make_pkg_dep_df}
\title{Make a package dependency dataframe from a list}
\usage{
make_pkg_dep_df(pkg_dependencies)
}
\arguments{
\item{pkg_dependencies}{a package dependency matrix frompkg_dependencies <- tools::package_dependencies tools::package_dependencies}
}
\value{
dataframe object with columns Package and Dependency, which are the 
package name and the name of one of its dependents respectively
}
\description{
Make a package dependency dataframe from a list
}
