% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_braces.R
\name{match_braces}
\alias{match_braces}
\title{Returns dataframe with ordered locations of the matching braces.}
\usage{
match_braces(txt)
}
\arguments{
\item{txt}{character vector of length one having 0 or more matching braces.}
}
\description{
Returns dataframe with ordered locations of the matching braces.
}
\examples{
library(rmsutilityr)
match_braces("{123{456{78}9}10}")
}
