% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_html_comment_lines.R
\name{merge_html_comment_lines}
\alias{merge_html_comment_lines}
\title{Merge multiline HTML comments into a single character string.}
\usage{
merge_html_comment_lines(lines, isolated = isolated)
}
\arguments{
\item{lines}{Character vector of lines to examine.}

\item{isolated}{Logical value indicating that the comment is the first text
to appear on the line.}
}
\description{
Merge multiline HTML comments into a single character string.
}
