% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{myDebug}
\alias{myDebug}
\title{Runs debug once on a function that is in memory}
\usage{
myDebug(f)
}
\arguments{
\item{f}{name of function to be debuged.}
}
\description{
myDebug() is by Andrea Span
}
