% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_installed.R
\name{not_installed}
\alias{not_installed}
\title{Check to see of a package is installed}
\usage{
not_installed(mypkg)
}
\arguments{
\item{mypkg}{is a character variable containing a package name}
}
\description{
This function takes a package name and sees if it is installed on the
present system.
}
\examples{
not_installed("stringi")
}
