% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Returns the words that describe the integer values of the number provided
as an argument.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{number}
}
\description{
Returns the words that describe the integer values of the number provided
as an argument.
}
