% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagebreak.R
\name{pagebreak}
\alias{pagebreak}
\title{pagebreak that works for PDF (LaTeX) and HTML output from \code{knitr}}
\usage{
pagebreak()
}
\value{
Returns either a LaTeX command, \code{\\newpage}, if 
\code{is_latex_output()} is \code{TRUE}, an HTML command,
\code{<div style="page-break-before: always;" />} if 
\code{is_html_output()} is \code{TRUE}, and \code{NULL} if 
neither is \code{TRUE}.
}
\description{
pagebreak that works for PDF (LaTeX) and HTML output from \code{knitr}
}
\examples{
pagebreak()
}
