% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcomment.R
\name{rcomment}
\alias{rcomment}
\title{Wrapper function around \code{colfmt()} to make the font red.}
\usage{
rcomment(x)
}
\arguments{
\item{x}{character string with comment}
}
\description{
In addition to adding color, underscores are prepended and
append for HTML and \code{\\emph{}} is used for LaTeX
}
