% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_cols.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Returns a dataframe after empty columns have been removed.}
\usage{
remove_empty_columns(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Returns a dataframe after empty columns have been removed.
}
