% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{remove_string}
\alias{remove_string}
\title{Returns a character vector with the remove_str pattern removed if it existed.
Warning: depricated; left in for compatibility. Use remove_strings() instead.}
\usage{
remove_string(strings, remove_str)
}
\arguments{
\item{strings}{character vector}

\item{remove_str}{character vector of length one containing the regular
expression of the character string to be removed from the strings character
vector}
}
\description{
Returns a character vector with the remove_str pattern removed if it existed.
Warning: depricated; left in for compatibility. Use remove_strings() instead.
}
