% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{remove_strings}
\alias{remove_strings}
\title{Returns a character vector with the string patterns within the expunge
character vector removed.}
\usage{
remove_strings(.str, expunge, ignore_case = FALSE)
}
\arguments{
\item{.str}{character vector}

\item{expunge}{character vector of strings to be removed}

\item{ignore_case}{logical indication whether or not to be case specific.}
}
\description{
Uses stri_detect_regex.
}
\examples{
strngs <- c("abc", "d", "E", "Fh")
expunge <- c("abc", "D", "E")
remove_strings(strngs, expunge, ignore_case = TRUE)
remove_strings(strngs, expunge, ignore_case = FALSE)

}
