% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_html_comment_text_lines_and_labels.R
\name{return_html_comment_text_lines_and_labels}
\alias{return_html_comment_text_lines_and_labels}
\title{Get labels, comments, and line numbers from HTML comments}
\usage{
return_html_comment_text_lines_and_labels(lines, label = "")
}
\arguments{
\item{lines}{Lines of text that may have comments.}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding each element of the character vector in turn 
immediately after "<!--" in the regex expression. The resulting logical
vectors are OR'd together to combine their results.}
}
\value{
List with `comment_lines`, which is an integer vector of line 
numbers indicating which lines are comments and `comment_labels`, which is 
a character vector with each comment label.
}
\description{
Get labels, comments, and line numbers from HTML comments
}
\examples{
return_html_comment_text_lines_and_labels(c("<!-- label1 comments -->", 
                                            "<!-- -->"))
return_html_comment_text_lines_and_labels(c("<!-- label1 comments -->", 
                                            "<!-- label2 comments-->"))
return_html_comment_text_lines_and_labels(c("<!-- label1 comments -->", 
                                            "<!-- label2 comments-->",
                                            "<!-- label3 comments -->"),
                                          label = "label1")
return_html_comment_text_lines_and_labels(c("<!-- label1 first comment -->", 
                                            "<!-- label2 second comment-->",
                                            "<!-- label3 third comment -->"),
                                          label = c("label1", "label3"))

}
