% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_detect_regex_all.R
\name{str_detect_regex_all}
\alias{str_detect_regex_all}
\title{Returns a logical vector with results of stri_detect() for each pattern in
second parameters character vector.}
\usage{
str_detect_regex_all(strings, patterns, ...)
}
\arguments{
\item{strings}{input vector. This must be an atomic vector and will be
coerced to a character vector.}

\item{patterns}{patterns to look for, as defined by a POSIX regular
expression. See the \emph{Extended Regular Expressions} section of regex
for details. See fixed, ignore.case and perl for how to use other
types of matching: fixed, case insensitive and perl-compatible.}

\item{...}{further arguments for stri_detect_regex}
}
\description{
Returns a logical vector with results of stri_detect() for each pattern in
second parameters character vector.
}
