% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_or_na_replace.R
\name{str_or_na_replace}
\alias{str_or_na_replace}
\title{Replaces string or NA in a character vector with either NA or
another string.}
\usage{
str_or_na_replace(x, pattern, replacement, fold = FALSE)
}
\arguments{
\item{x}{character vector}

\item{pattern}{string pattern to search for may be empty string or NA.
Comparison is done with "==". Case can be ignored.}

\item{replacement}{string to replace entire entity that has pattern.}

\item{fold}{logical value where if TRUE the vector \code{x} and
\code{pattern} are converted to lower case.}
}
\description{
This is to augment \code{stri_replace_all_fixed()}, which does
not allow \code{NA} as a pattern.
}
