% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{str_reverse}
\alias{str_reverse}
\title{Returns character vector where each element has had its character
order rerversed.}
\usage{
str_reverse(x)
}
\arguments{
\item{x}{- Character vector}
}
\description{
Returns character vector where each element has had its character
order rerversed.
}
\examples{
str_reverse('abc') # "cba"
str_reverse(c('abc', 'defgh', 'i', 'jklm')) # "cba" "hgfed" "i" "mlkj"
}
