% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stri_trim_cols.R
\name{stri_trim_cols}
\alias{stri_trim_cols}
\title{Takes a dataframe and trims all character columns of leading and trailing
blanks.}
\usage{
stri_trim_cols(.df, cols = names(.df))
}
\arguments{
\item{.df}{dataframe to have columns trimmed}

\item{cols}{optional character vector arguement specifying which columns to
trim}
}
\value{
dataframe with values in all character columns having been
trimmed on both sides.
}
\description{
Takes a dataframe and trims all character columns of leading and trailing
blanks.
}
