% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{underscoreToTitleCase}
\alias{underscoreToTitleCase}
\title{Convert underscore text strings such as \code{trans_col_select} to
\code{TransColSelect}.}
\usage{
underscoreToTitleCase(s)
}
\arguments{
\item{s}{character vector that may have tokens with embedded
underscores.}
}
\description{
Convert underscore text strings such as \code{trans_col_select} to
\code{TransColSelect}.
}
