% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_files_after_deleting_selected_comments.R
\name{write_files_after_deleting_selected_comments}
\alias{write_files_after_deleting_selected_comments}
\title{Write files after deleting selected comments.}
\usage{
write_files_after_deleting_selected_comments(
  new_path,
  files,
  label = "",
  isolated = TRUE
)
}
\arguments{
\item{new_path}{Character vector of length one indicating the directory
newly created files are to be placed.}

\item{files}{Character vector of files to parse.}

\item{label}{Optional regex expression that can be used to limit the 
comments found by adding each element of the character vector in turn 
immediately after "<!--" in the regex expression. The resulting logical
vectors are OR'd together to combine their results.}

\item{isolated}{Logical value indicating that the comment is the first text
to appear on the line.}
}
\description{
Write files after deleting selected comments.
}
