% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt_print.R
\name{xt_print}
\alias{xt_print}
\title{Prints xtable with gray bar}
\usage{
xt_print(df, caption = "", label = "", size = "normalsize",
  cont = "Cont. ...", tabular.environment = "longtable",
  include.rownames = FALSE, align = NULL, floating = FALSE,
  floating.environment = "table", digits = NULL, hlines.after = c(-1, 0,
  nrow(df)), type = "latex", ...)
}
\arguments{
\item{df}{data frame being printed}

\item{caption}{table caption. Default is ''.}

\item{label}{table label. Default is ''.}

\item{size}{latex font size to be used. Default is 'normalsize'}

\item{cont}{longtable continuation string. Default is 'Cont. ...'.}

\item{tabular.environment}{print.xtable tabular.environment. Default is
'longtable'.}

\item{include.rownames}{logical. If TRUE the rows names is printed.
Default value is TRUE.}

\item{align}{Character vector of length equal to the number of columns
of the resulting table, indicating the alignment of the corresponding
columns. Default is NULL
Also, "|" may be used to produce vertical lines between columns in
LaTeX tables, but these are effectively ignored when considering the
required length of the supplied vector. If a character vector of length
one is supplied, it is split as strsplit(align, "")[[1]] before processing.
Since the row names are printed in the first column, the length of align
is one greater than ncol(x) if x is a data.frame. Use "l", "r", and "c"
to denote left, right, and center alignment, respectively. Use "p{3cm}" etc.
for a LaTeX column of the specified width. For HTML output the "p" alignment
is interpreted as "l", ignoring the width request.}

\item{floating}{If TRUE and type="latex", the resulting table will be a
floating table (using, for example, \\begin{table} and \\end{table}).
See floating.environment below. Default value is TRUE.}

\item{floating.environment}{If floating=TRUE and type="latex",
the resulting table uses the specified floating environment.
Possible values include "table", "table*", and other floating
environments defined in LaTeX packages. Default value is "table".}

\item{digits}{Numeric vector of length equal to one (in which case
it will be replicated as necessary) or to the number of columns of the
resulting table or matrix of the same size as the resulting table
indicating the number of digits to display in the corresponding columns.
Since the row names are printed in the first column, the length of the
vector digits or the number of columns of the matrix digits is one
greater than ncol(x) if x is a data.frame. Default depends of class of x.
If values of digits are negative, the corresponding values of x are
displayed in scientific format with abs(digits) digits.}

\item{hlines.after}{When type="latex", a vector of numbers between -1 and
"nrow(x)", inclusive, indicating the rows after which a horizontal line
should appear. If NULL is used no lines are produced. Default value is
c(-1,0,nrow(x)) which means draw a line before and after the columns
names and at the end of the table. Repeated values are allowed.}

\item{type}{character vector of length one representing the type of table
to produce. Possible values for type are "latex" or "html".
Default value is "latex".}

\item{...}{extra arguments to print()}
}
\description{
Prints xtable with longtable features
}
