% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsutilityr.R
\name{get_deps}
\alias{get_deps}
\title{Has the sideeffect of writing out a LaTeX title page compatible with memoir.}
\usage{
get_deps(path)
}
\arguments{
\item{path}{character vector of length one having the directory path to
where new version of the custom package exists.}
}
\description{
#' Written by Josh O'Brien on stackoverflow on May 13 '15 at 21:42
}
