% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_descriptions.R
\name{get_pkg_descriptions}
\alias{get_pkg_descriptions}
\title{Provides a dataframe of package description fields}
\usage{
get_pkg_descriptions(pkgs = NULL, lib.loc = NULL,
  fields = c("Package", "Type", "Title", "Date", "Author", "Creator",
  "Maintainer", "Description", "Depends", "Imports", "Suggests",
  "Encoding", "License", "RoxygenNote", "LazyData", "VignetteBuilder",
  "URL", "BugReports"), base = FALSE, dependencies = FALSE,
  which = c("Depends", "Imports", "LinkingTo"), recursive = TRUE,
  reverse = FALSE, verbose = getOption("verbose"))
}
\arguments{
\item{pkgs}{character vector of package names. Defaults to packages found with
sessionInfo().}

\item{lib.loc}{a character vector of directory names of R libraries, or NULL. 
The default value of NULL corresponds to all libraries currently known. 
If the default is used, the loaded packages and namespaces are searched 
before the libraries.}

\item{fields}{a character vector giving the tags of fields to return 
(if other fields occur in the file they are ignored).}

\item{base}{logical value if TRUE base packages are included}

\item{dependencies}{logical value if TRUE, dependencies are recursively 
added.}

\item{which}{a character vector listing the types of dependencies, 
a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances").
Character string "all" is shorthand for that vector, character string 
"most" for the same vector without "Enhances".}

\item{recursive}{logical: should (reverse) dependencies of (reverse)
dependencies (and so on) be included?}

\item{reverse}{logical: if FALSE (default), regular dependencies are 
calculated, otherwise reverse dependencies.}

\item{verbose}{logical indicating if output should monitor the 
package search cycles.}
}
\value{
dataframe with p
}
\description{
Provides a dataframe of package description fields
}
\examples{
write.csv(data.frame(packages = get_pkg_list(base = TRUE), stringsAsFactors = FALSE), 
          file = "package_list.csv", row.names = FALSE, quote = FALSE)

}
