% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Returns the words that describe the integer values of the number provided
as an argument.}
\usage{
numbers2words(x)
}
\arguments{
\item{x}{scaler that may be integer or number (real).}
}
\description{
From GitHubGistpsychemedia/numbers2words.R
Function by John Fox found here:
http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
Tweaks by AJH to add commas and "and" have been commented out since they
are wrong
}
